Fixed a long standing warning
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@27819 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/common/utility/MRUCache.h
@@ -497,8 +497,8 @@ template<class kT, class vT, MRUCache_DeleteMethod dM, size_t nS>
 kT MRUCacheBase<kT,vT,dM,nS>::oldest(void)
 {
    typename std::map<kT,size_t>::iterator i = age.begin();
-   kT retval               = i->first;
-   int oldestAge           = i->second; 
+   kT retval         = i->first;
+   size_t oldestAge  = i->second; 
 
    for (i = age.begin(); i != age.end(); i++)
    {
